initial reorganization implementation
