+++ evaltools/data/AssignmentCompressor.py
