+++ evaltools/scores/__init__.py
@@ -6,6 +6,7 @@ Basic functionality for evaluating districting plans.
 from .splits import splits, pieces
 from .population import deviations, unassigned_population
 from .contiguity import unassigned_units, contiguous
+from .reock import reock
 
 __all__ = [
     "splits",
@@ -13,5 +14,6 @@ __all__ = [
     "deviations",
     "unassigned_population",
     "unassigned_units",
-    "contiguous"
+    "contiguous",
+    "reock"
 ]
