+++ tests/test_geography.py
@@ -1,7 +1,5 @@
 
-from evaltools.geography import dissolve, dualgraph, unitmap, invert, reock
-from gerrychain.grid import Grid
-from shapely.geometry import box
+from evaltools.geometry import dissolve, dualgraph, unitmap, invert
 import geopandas as gpd
 from pathlib import Path
 import os
@@ -76,60 +74,6 @@ def test_unitmap():
     # Assert that we have a dict and that it has as many keys as counties.
     assert type(inverse) is dict
     assert len(inverse) == len(counties)
-
-
-def test_reock_score_squares_geodataframe():
-    grid = Grid((10, 10))
-    gdf = gpd.GeoDataFrame([
-        {'node': (x, y), 'geometry': box(x, y, x + 1, y + 1)}
-        for (x, y) in grid.graph
-    ]).set_index('node')
-    score_fn = reock(gdf)
-
-    # The Reock score of a square inscribed in a circle is
-    # area(square) / area(circle) = 2/π.
-    expected_dist_score = 2/ math.pi
-    scored = score_fn(grid)
-
-    assert scored.keys() == grid.parts.keys()
-    for dist_score in scored.values():
-        assert abs(dist_score - expected_dist_score) < 1e-4
-
-
-def test_reock_score_squares_graph():
-    grid = Grid((10, 10))
-    for (x, y), data in grid.graph.nodes(data=True):
-        data['geometry'] = box(x, y, x + 1, y + 1)
-    score_fn = reock(grid.graph)
-
-    # The Reock score of a square inscribed in a circle is
-    # area(square) / area(circle) = 2/π.
-    expected_dist_score = 2 / math.pi
-    scored = score_fn(grid)
-
-    assert scored.keys() == grid.parts.keys()
-    for dist_score in scored.values():
-        assert abs(dist_score - expected_dist_score) < 1e-4
-
-
-def test_reock_score_disconnected():
-    grid = Grid((10, 10))
-    for (x, y), data in grid.graph.nodes(data=True):
-        data['geometry'] = box(x, y, x + 1, y + 1)
-    score_fn = reock(grid.graph)
-
-    # Break district 0 into two disconnected pieces 
-    # (while preserving convex hull perimeter),
-    # adding a disconnected component to district 3
-    # (quadrupling convex hull perimeter).
-    grid_disconnected = grid.flip({(0, 1): 3, (1, 0): 3})
-
-    scored = score_fn(grid_disconnected)
-    assert scored.keys() == grid_disconnected.parts.keys()
-    assert abs(scored[0] - ((23 / 25) * (2 / math.pi))) < 1e-4
-    assert abs(scored[1] - (2 / math.pi)) < 1e-4
-    assert abs(scored[2] - (2 / math.pi)) < 1e-4
-    assert abs(scored[3] == (27 / 25) * (math.pi / 2)) < 1e-4
 
 if __name__ == "__main__":
     root = Path(os.getcwd()) / Path("test-resources/")
